update source files to remove unused snippets
