+++ examples/dart/test/define_realm_model_test.dart
@@ -1,21 +1,12 @@
-// Used in quickstart
-// :snippet-start: define-model
-// :state-start: dart
 import 'package:realm_dart/realm.dart';
 import 'package:test/expect.dart';
 import 'package:test/scaffolding.dart';
-// :state-end:
 
-// :state-start: flutter
-// :state-uncomment-start: flutter
-// import 'package:realm/realm.dart';
-// :state-uncomment-end:
-// :state-end:
+part 'define_realm_model_test.g.dart';
 
-part 'define_realm_model_test.g.dart'; // :remove:
-// :uncomment-start:
-// part 'car.g.dart';
-// :uncomment-end:
+// TODO: the snippet source/examples/generated/flutter/define_realm_model_test.snippet.define-model.dart
+// doesn't come from the source code. there was a bit of hackery making it so.
+// fix to create.
 
 @RealmModel()
 class _Car {
@@ -26,7 +17,6 @@ class _Car {
   late String? model;
   late int? miles;
 }
-// :snippet-end:
 
 // :snippet-start: modify-schema-model
 @RealmModel()
